+++ node/commons/compatibility-layer.ts
@@ -3,6 +3,10 @@ import unescape from 'unescape'
 import { Checkout } from '../clients/checkout'
 import { groupBy, prop, indexBy, mergeAll } from 'ramda'
 import { removeDiacriticsFromURL } from '../utils/string'
+import {
+  formatTranslatableStringV2
+} from '@vtex/api'
+
 
 export enum IndexingType {
   API = 'API',
@@ -24,6 +28,7 @@ export const convertBiggyProduct = async (
   tradePolicy?: string,
   priceTable?: string,
   regionId?: string,
+  locale?: string,
   indexingType?: IndexingType,
 ) => {
   const categories: string[] = []
@@ -35,7 +40,7 @@ export const convertBiggyProduct = async (
   })
 
   const skus: SearchItem[] = (product.skus || []).map(
-    convertSKU(product, indexingType, tradePolicy)
+    convertSKU(product, indexingType, tradePolicy, locale)
   )
 
   const allSpecifications = (product.productSpecifications ?? []).concat(getSKUSpecifications(product))
@@ -58,13 +63,25 @@ export const convertBiggyProduct = async (
     categoriesIds,
     productId: product.id,
     cacheId: `sp-${product.id}`,
-    productName: product.name,
+    productName: formatTranslatableStringV2({
+      content: product.name,
+      context: product.id,
+      from: locale,
+    }),
     productReference: product.reference || product.product || product.id,
     linkText: product.link,
-    brand: product.brand || '',
+    brand: formatTranslatableStringV2({
+      content: product.brand || '',
+      context: product.id,
+      from: locale,
+    }),
     brandId,
     link: product.url,
-    description: product.description,
+    description: formatTranslatableStringV2({
+      content: product.description,
+      context: product.id,
+      from: locale,
+    }),
     items: skus,
     allSpecifications,
     categoryId: product.categoryIds?.slice(-1)[0],
@@ -352,7 +369,8 @@ const convertImages = (images: ElasticImage[], indexingType?: IndexingType) => {
 const convertSKU = (
   product: BiggySearchProduct,
   indexingType?: IndexingType,
-  tradePolicy?: string
+  tradePolicy?: string,
+  locale?: string,
 ) => (sku: BiggySearchSKU): SearchItem & { [key: string]: any } => {
   const images = convertImages(product.images, indexingType)
 
@@ -367,9 +385,21 @@ const convertSKU = (
     sellers,
     images,
     itemId: sku.id,
-    name: product.name,
-    nameComplete: product.name,
-    complementName: product.name,
+    name: formatTranslatableStringV2({
+      content: product.name,
+      context: sku.id,
+      from: locale,
+    }),
+    nameComplete: formatTranslatableStringV2({
+      content: product.name,
+      context: sku.id,
+      from: locale,
+    }),
+    complementName: formatTranslatableStringV2({
+      content: product.name,
+      context: sku.id,
+      from: locale,
+    }),
     referenceId: [
       {
         Key: 'RefId',